adding token on server

+++ api/resource/user_resource.py
@@ -4,6 +4,9 @@ from flask import Blueprint, request
 from repository.user_repository import UserRepository
 from shared.utils import default_parser
 
+from itsdangerous import (TimedJSONWebSignatureSerializer
+                          as Serializer, BadSignature, SignatureExpired)
+
 import simplejson as json
 ''' This class recive http requests from web clients to retrive, save and update data
 '''
@@ -12,33 +15,62 @@ user_repository = UserRepository()
 
 user_blueprint = Blueprint('user_blueprint', __name__)
 
-@user_blueprint.route("/users", methods=['GET'])
-def list():
+def check_token(token):
+    return user_repository.checkAuth(token)
+
+@user_blueprint.route("/users/<string:token>", methods=['GET'])
+def list(token):
     '''Get a list of all users
     '''
-    return json.dumps(user_repository.list(), default=default_parser), 200
+    if not check_token(token):
+        return 'Invalid token', 400
 
+    try:
+        return json.dumps(user_repository.list(), default=default_parser), 200
+    except Exception as e:
+        return '%s' % (e), 400
 
-@user_blueprint.route("/user/<string:id>", methods=['GET'])
-def get(id):
+@user_blueprint.route("/user/<string:id>/<string:token>", methods=['GET'])
+def get(id, token):
     '''Get a specific user
     '''
-    user = user_repository.get(id)
-    return json.dumps(user, default=default_parser), 200
 
-@user_blueprint.route("/list-emails", methods=['GET'])
-def get_registered_emails():
+    if not check_token(token):
+        return 'Invalid token', 400
+
+    try:
+        user = user_repository.get(id)
+        return json.dumps(user, default=default_parser), 200
+    except Exception as e:
+        return '%s' % (e), 400
+
+@user_blueprint.route("/list-emails/<string:token>", methods=['GET'])
+def get_registered_emails(token):
     ''' Return a list of users emails
     '''
-    list_of_emails = user_repository.get_all_emails()
-    return json.dumps(list_of_emails, default=default_parser), 200
 
-@user_blueprint.route("/email/<string:id>", methods=['GET'])
-def get_email_id(id):
+    if not check_token(token):
+        return 'Invalid token', 400
+
+    try:
+        list_of_emails = user_repository.get_all_emails()
+        return json.dumps(list_of_emails, default=default_parser), 200
+    except Exception as e:
+        return '%s' % (e), 400
+
+@user_blueprint.route("/email/<string:id>/<string:token>", methods=['GET'])
+def get_email_id(id, token):
     '''Get email of a specific user
     '''
-    email = user_repository.get_email(id)
-    return json.dumps(email, default=default_parser), 200
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
+    try:
+        email = user_repository.get_email(id)
+        return json.dumps(email, default=default_parser), 200
+    except Exception as e:
+        return '%s' % (e), 400
 
 @user_blueprint.route("/user/<string:email>/<string:password>", methods=['GET'])
 def autenticate(email, password):
@@ -67,10 +99,14 @@ def post():
         return '%s' % (e), 400
 
 
-@user_blueprint.route("/user", methods=['PUT'])
-def update():
+@user_blueprint.route("/user/<string:token>", methods=['PUT'])
+def update(token):
     '''Update user informations
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     user_repository.update(
         json.loads(
             request.data.decode('utf-8')
@@ -79,19 +115,30 @@ def update():
     return "User successfully updated", 200
 
 
-@user_blueprint.route("/user/<string:id>", methods=['DELETE'])
-def delete(id):
+@user_blueprint.route("/user/<string:id>/<string:token>", methods=['DELETE'])
+def delete(id, token):
     '''Let the clients delete a user
     '''
-    user_repository.delete(id)
-    return 'User successfully deleted', 200
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
+    try:
+        user_repository.delete(id)
+        return 'User successfully deleted', 200
+    except Exception as e:
+        return '%s' % (e), 400
 
 # Documents and folders!
 
-@user_blueprint.route("/document/<string:user_id>/<string:folder_id>", methods=['POST'])
-def new_document(user_id, folder_id):
+@user_blueprint.route("/document/<string:user_id>/<string:folder_id>/<string:token>", methods=['POST'])
+def new_document(user_id, folder_id, token):
     '''Let the clients create a document to an user saved documents
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         kwargs = json.loads(request.data.decode('utf-8'))
         kwargs['user_id'] = user_id
@@ -101,10 +148,14 @@ def new_document(user_id, folder_id):
     except Exception as e:
         return '%s' % (e), 404
 
-@user_blueprint.route("/document/<string:user_id>", methods=['DELETE'])
-def delete_document(user_id):
+@user_blueprint.route("/document/<string:user_id>/<string:token>", methods=['DELETE'])
+def delete_document(user_id, token):
     ''' Let the clients delete a documet from an user
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         kwargs = json.loads(request.data.decode('utf-8'))
         kwargs['user_id'] = user_id
@@ -113,10 +164,14 @@ def delete_document(user_id):
     except Exception as e:
         return '%s' % (e), 404
 
-@user_blueprint.route("/document/<string:user_id>", methods=['PUT'])
-def edit_document(user_id):
+@user_blueprint.route("/document/<string:user_id>/<string:token>", methods=['PUT'])
+def edit_document(user_id, token):
     ''' Edit a user document, update the document content
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         kwargs = json.loads(request.data.decode('utf-8'))
         kwargs['user_id'] = user_id
@@ -126,10 +181,14 @@ def edit_document(user_id):
     except Exception as e:
         return '%s' % (e), 404
 
-@user_blueprint.route("/share/edit", methods=['PUT'])
-def edit_document_shared():
+@user_blueprint.route("/share/edit/<string:token>", methods=['PUT'])
+def edit_document_shared(token):
     ''' Let the documents edit the content of a user shared document
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         kwargs = json.loads(request.data.decode('utf-8'))
         print kwargs
@@ -139,10 +198,14 @@ def edit_document_shared():
     except Exception as e:
         return '%s' % (e), 404
 
-@user_blueprint.route("/folders/<string:user_id>", methods=['GET'])
-def get_folders(user_id):
+@user_blueprint.route("/folders/<string:user_id>/<string:token>", methods=['GET'])
+def get_folders(user_id, token):
     ''' Retrive all user folders
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         folders = user_repository.get_folders(user_id)
         return json.dumps(folders, default=default_parser), 200
@@ -151,20 +214,28 @@ def get_folders(user_id):
 
 
 
-@user_blueprint.route("/folders_tree/<string:user_id>", methods=['GET'])
-def get_folders_tree(user_id):
+@user_blueprint.route("/folders_tree/<string:user_id>/<string:token>", methods=['GET'])
+def get_folders_tree(user_id, token):
     '''Retrive all folders and documents on a node tree format to let cliets show the document in a more friendly way
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         folders = [user_repository.get_folders_tree(user_id)]
         return json.dumps(folders, default=default_parser), 200
     except Exception as e:
         return '%s' % (e), 404
 
-@user_blueprint.route("/folder/<string:user_id>", methods=['POST'])
-def new_folder(user_id):
+@user_blueprint.route("/folder/<string:user_id>/<string:token>", methods=['POST'])
+def new_folder(user_id, token):
     ''' Create a folder to a user
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         kwargs = json.loads(request.data.decode('utf-8'))
         kwargs['user_id'] = user_id
@@ -173,10 +244,14 @@ def new_folder(user_id):
     except Exception as e:
         return "%s" % (e), 404
 
-@user_blueprint.route("/folder/<string:user_id>/<string:folder_id>", methods=['PUT'])
-def rename_Folder(user_id, folder_id):
+@user_blueprint.route("/folder/<string:user_id>/<string:folder_id>/<string:token>", methods=['PUT'])
+def rename_Folder(user_id, folder_id, token):
     '''Rename a folder from user
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         kwargs = json.loads(request.data.decode('utf-8'))
         kwargs['user_id'] = user_id
@@ -186,10 +261,14 @@ def rename_Folder(user_id, folder_id):
     except Exception as e:
         return '%s' % (e), 404
 
-@user_blueprint.route("/folder/<string:user_id>/<string:folder_id>", methods=['DELETE'])
-def delete_folder(user_id, folder_id):
+@user_blueprint.route("/folder/<string:user_id>/<string:folder_id>/<string:token>", methods=['DELETE'])
+def delete_folder(user_id, folder_id, token):
     ''' Let the clients delete an user folder
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         kwargs = {}
         kwargs['user_id'] = user_id
@@ -200,10 +279,14 @@ def delete_folder(user_id, folder_id):
         return '%s' % (e), 404
 
 
-@user_blueprint.route("/share/<string:user_id>/<string:document_id>", methods=['PUT'])
-def share_document(user_id, document_id):
+@user_blueprint.route("/share/<string:user_id>/<string:document_id>/<string:token>", methods=['PUT'])
+def share_document(user_id, document_id, token):
     '''Share a document from an user with other user
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         kwargs = json.loads(request.data.decode('utf-8'))
         kwargs['user_id'] = user_id
@@ -214,10 +297,14 @@ def share_document(user_id, document_id):
     except Exception as e:
         return '%s' % (e), 404
 
-@user_blueprint.route("/share/<string:user_id>", methods=['GET'])
-def get_shared_documents(user_id):
+@user_blueprint.route("/share/<string:user_id>/<string:token>", methods=['GET'])
+def get_shared_documents(user_id, token):
     ''' Retrive all documets shareds with an user
     '''
+
+    if not check_token(token):
+        return 'Invalid token', 400
+
     try:
         shared_documents = user_repository.get_shared_with_me_documents(user_id)
         return json.dumps(shared_documents, default=default_parser), 200
